requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0

# # Deep Learning
# torch>=1.9.0
# stable-baselines3>=1.0

# MetaTrader connection
pyzmq>=22.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Utilities
pyyaml>=6.0
python-dotenv>=0.19.0
click>=8.0.0
tqdm>=4.62.0
MetaTrader5